# packages/contracts/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AEP BFF API
  version: 1.0.0
servers:
- url: /api/v1
paths:
...
  /projects/{projectName}/agents/{conversationId}/messages:
    get:
      operationId: get-conversation
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: FE-chosen conversation uuid
        in: path
        name: conversationId
        required: true
        schema:
          description: FE-chosen conversation uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Rehydrate a chat conversation (messages only)
      tags:
      - GenAI
    post:
      operationId: create-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: FE-chosen conversation uuid
        in: path
        name: conversationId
        required: true
        schema:
          description: FE-chosen conversation uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TurnInputBody'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnOutputBody'
          description: Accepted
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnConflict'
          description: A turn is already running for this project, or the design
            use-case has no requirements yet.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Start a generation/chat turn (202 — runs detached; attach via the turn stream)
      tags:
      - GenAI
...
  /projects/{projectName}/turns/active:
    get:
      operationId: get-active-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnStatus'
          description: OK
        '204':
          description: No running turn
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Get the project's running turn (204 when none)
      tags:
      - GenAI
  /projects/{projectName}/turns/{turnId}:
    get:
      operationId: get-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Turn id from the 202 create response
        in: path
        name: turnId
        required: true
        schema:
          description: Turn id from the 202 create response
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnStatus'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Get one turn's status
      tags:
      - GenAI
  /projects/{projectName}/turns/{turnId}/stream:
    get:
      operationId: stream-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Turn id from the 202 create response
        in: path
        name: turnId
        required: true
        schema:
          description: Turn id from the 202 create response
          type: string
      - description: Replay from this absolute event index (wins over Last-Event-ID)
        in: query
        name: from
        schema:
          # Keep an explicit pointer under the global prefer-skip-optional-pointer:
          # the handler must distinguish 'absent' from the zero value (from).
          x-go-type-skip-optional-pointer: false
          description: Replay from this absolute event index (wins over Last-Event-ID)
          minimum: -1
          type: integer
      - description: 'SSE auto-reconnect resume cursor: the last frame id the client saw; replay
          resumes after it. Opaque per the SSE spec — non-numeric values are ignored (full replay).
          The from query param wins when both are present.'
        in: header
        name: Last-Event-ID
        schema:
          type: string
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                description: Raw agent StreamPart frames + one terminal event, id-stamped; ends with [DONE]
                type: string
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Attach to a turn's event stream (SSE replay + live tail)
      tags:
      - GenAI
...
components:
  schemas:
...
    Error:
      description: Flat error envelope returned by every non-2xx response.
      type: object
      additionalProperties: false
      properties:
        code:
          description: Machine-readable error code (stable slug).
          type: string
          example: validation_failed
        message:
          description: Human-readable error message.
          type: string
          example: request body is invalid
        details:
          description: Field-level details for validation errors.
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required:
      - code
      - message
    ErrorDetail:
      description: One field-level validation error.
      type: object
      additionalProperties: false
      properties:
        field:
          description: Location of the offending field (e.g. body.name).
          type: string
        message:
          description: What is wrong with the field.
          type: string
      required:
      - field
      - message
...
    TurnInputBody:
      additionalProperties: false
      properties:
        collab:
          description: 'Room-scoped turn (#86 phase 4): the agent joins the project''s spec collab room
            as a live peer, reads and edits the shared doc, and commits nothing to git.'
          type: boolean
        instruction:
          description: User message / generation directive
          type: string
        target:
          description: Optional target (e.g. a doc type)
          type: string
        useCase:
          description: Which generation/chat flow. Omit to run a generic spec turn.
          enum:
          - requirements-generate
          - requirements-chat
          - design-generate
          type: string
      required:
      - instruction
      type: object
    TurnOutputBody:
      additionalProperties: false
      properties:
        turnId:
          description: The started turn's id — poll/attach with it
          type: string
      required:
      - turnId
      type: object
    TurnStatus:
      additionalProperties: false
      description: One turn's lifecycle view (create-turn 202 → poll/attach).
      properties:
        commitSha:
          type: string
        conversationId:
          type: string
        createdAt:
          format: date-time
          type: string
        message:
          type: string
        noChanges:
          type: boolean
        paths:
          items:
            type: string
          type: array
          nullable: true
        reason:
          type: string
        status:
          description: running, completed, failed
          type: string
        turnId:
          type: string
        updatedAt:
          format: date-time
          type: string
        useCase:
          type: string
      required:
      - turnId
      - conversationId
      - useCase
      - status
      - createdAt
      - updatedAt
      type: object
...
    TurnConflict:
      additionalProperties: false
      description: create-turn 409 body. turn_in_progress carries the active
        turn's id; requirements_missing means the design use-case has no
        requirements to work from.
      properties:
        code:
          enum:
          - turn_in_progress
          - requirements_missing
          type: string
        activeTurnId:
          type: string
      required:
      - code
      type: object
...
  securitySchemes:
    userJWT:
      bearerFormat: JWT
      description: Thunder end-user JWT (RS256, JWKS-verified)
      scheme: bearer
      type: http